Change some ORs to plus
Port from snappy SVN

Change a few ORs to additions where they don't matter. This helps the
compiler use the LEA instruction more efficiently, since e.g. a + (b << 2)
can be encoded as one instruction. Even more importantly, it can constant-fold the
COPY_* enums together with the shifted negative constants, which
also saves some instructions. (We don't need it for LITERAL, since it happens
to be 0.)

I am unsure why the compiler couldn't do this itself, but the theory
is that it cannot prove that len-1 and len-4 cannot underflow/wrap, and thus
can't do the optimization safely.